src/Defra.PTS.Web.Domain/ViewModels/TravelDocument/PetNameViewModel.cs
using Defra.PTS.Web.Domain.Enums;

namespace Defra.PTS.Web.Domain.ViewModels.TravelDocument;

public class PetNameViewModel : TravelDocumentFormPage
{
    public string FormTitle => $"What is your pet's name?";

    public string PetName { get; set; }

    public override Enums.TravelDocumentFormPageType PageType => Enums.TravelDocumentFormPageType.PetName;

    public override void TrimUnwantedData()
    {
    }

    public override void ClearData()
    {
        IsCompleted = default;
        PetName = default;
    }
}
